utworzenie klasy HibernateUtil

+++ mavenstart/src/test/java/com/ownProjects/webLangSite/HelloServiceTest.java
@@ -15,7 +15,7 @@ private final static String WELCOME="Hello";
         //given
         var mockRepository=new LangRepository(){
             @Override
-            Optional<Lang> findById(Long id) {
+            Optional<Lang> findById(Integer id) {
                 return Optional.of(new Lang(null,WELCOME,null));
             }
         };
@@ -30,7 +30,7 @@ private final static String WELCOME="Hello";
         //given
         var mockRepository=new LangRepository(){
             @Override
-            Optional<Lang> findById(Long id) {
+            Optional<Lang> findById(Integer id) {
                 return Optional.of(new Lang(null,WELCOME,null));
             }
         };
@@ -47,7 +47,7 @@ private final static String WELCOME="Hello";
         var defaultIdWelcome="Ciao";
         var mockRepository=new LangRepository(){
             @Override
-            Optional<Lang> findById(Long id) {
+            Optional<Lang> findById(Integer id) {
                 if (id.equals(HelloService.DEFAULT_LANG.getId())){
                     return Optional.of(new Lang(null,defaultIdWelcome,null));
                 }
@@ -67,7 +67,7 @@ private final static String WELCOME="Hello";
         var defaultIdWelcome="Ciao";
         var mockRepository=new LangRepository(){
             @Override
-            Optional<Lang> findById(Long id) {
+            Optional<Lang> findById(Integer id) {
                 if (id.equals(HelloService.DEFAULT_LANG.getId())){
                     return Optional.empty();
                 }
@@ -87,7 +87,7 @@ private final static String WELCOME="Hello";
         var defaultIdWelcome="Ciao";
         var mockRepository=new LangRepository(){
             @Override
-            Optional<Lang> findById(Long id) {
+            Optional<Lang> findById(Integer id) {
                 if (id.equals(HelloService.DEFAULT_LANG.getId())){
                     return Optional.of(new Lang(null,defaultIdWelcome,null));
                 }
